+++ src/app/shared/category.service.ts
@@ -1,8 +1,9 @@
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
-import { Subject } from "rxjs";
+import { of, Subject } from "rxjs";
 import { Observable } from "rxjs-compat";
 import { Category } from "./category.model";
+import { environment } from "../../environments/environment";
 
 
 @Injectable()
@@ -11,8 +12,18 @@ export class CategoryService {
     private apiUrl = 'http://127.0.0.1:8000/api/categories/';
 
     private categoriesChanged = new Subject<void>();
-
     categoriesChanged$ = this.categoriesChanged.asObservable();
+
+    private mockCategories: Category[] = [
+        { id: 1, name: "Air Fryer"},
+        { id: 1, name: "Bread"},
+        { id: 1, name: "Chicken"},
+        { id: 1, name: "Beef"},
+        { id: 1, name: "Pasta"},
+        { id: 1, name: "Sauce"},
+        { id: 1, name: "Instant Pot"},
+        { id: 1, name: "Salad"},
+    ]
 
     notifyCategoriesChanged() {
         this.categoriesChanged.next();
@@ -22,17 +33,35 @@ export class CategoryService {
 
     //Get all categories
     getCategories(): Observable<Category[]> {
-        return this.http.get<Category[]>(this.apiUrl)
+        if(environment.useBackend){
+            return this.http.get<Category[]>(this.apiUrl);
+        } else{
+            return of(this.mockCategories);
+        }
     }
 
     //delete a specific category
     deleteCategory(id: number): Observable<void> {
-        return this.http.delete<void>(`${this.apiUrl}${id}/delete/`)
+        if (environment.useBackend){
+            return this.http.delete<void>(`${this.apiUrl}${id}/delete/`)
+        } else {
+            this.mockCategories = this.mockCategories.filter(cat => cat.id !== id);
+            this.notifyCategoriesChanged();
+            return of();
+        }
+        
     }
 
     //save new categories
     saveCategories(categories: Category[]): Observable<Category[]> {
-        return this.http.put<Category[]>(`${this.apiUrl}update/`, categories);
+        if(environment.useBackend){
+            return this.http.put<Category[]>(`${this.apiUrl}update/`, categories);
+        } else {
+            this.mockCategories = categories;
+            this.notifyCategoriesChanged();
+            return of(this.mockCategories);
+        }
+        
     }
 
 }
